backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.9
requests==2.32.3
supabase
httpx[http2]